+++ sw-sysemu/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Started development for 0.12.0
 - (Conan) Generate packages for linux-ubuntu22.04-x86_64-gcc11-release
 - [SW-17595] Auto-attach gdbserver on user mode code.
+- [SW-17597] Dump Trace-buffer on sysemu Fatals.
 ### Changed
 - (CI) Update gitlab-ci-common pointer for Gitlab v15 compatibility
 - Updated system-sw embedded components